refactor : Create lotto.core package
move Lotto, LottoGenerator, LottoNumber, LottoValidator, Lottos,
PurchaseAmount class to lotto.core package

test files corresponding to the above were also moved the lotto.core
package

Please enter the commit message for your changes. Lines starting
