+++ src/main/java/lotto/core/Lottos.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 import java.util.List;
 import java.util.Objects;
