+++ src/main/java/lotto/core/PurchaseAmount.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.core;
 
 import constants.ErrorMessage;
 import java.util.Objects;
