+++ src/test/java/lotto/core/LottoGeneratorTest.java
@@ -1,11 +1,11 @@
-package lotto;
+package lotto.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import lotto.Lotto;
-import lotto.LottoGenerator;
-import lotto.LottoNumber;
+import lotto.core.Lotto;
+import lotto.core.LottoGenerator;
+import lotto.core.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 public class LottoGeneratorTest {
