+++ src/test/java/lotto/core/PurchaseAmountTest.java
@@ -1,10 +1,10 @@
-package lotto;
+package lotto.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
-import lotto.PurchaseAmount;
+import lotto.core.PurchaseAmount;
 import org.junit.jupiter.api.Test;
 
 public class PurchaseAmountTest {
