+++ src/test/java/lotto/statistics/RevenueCalculatorTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.within;
 
 import java.util.Arrays;
-import lotto.PurchaseAmount;
+import lotto.core.PurchaseAmount;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
