+++ src/test/java/lotto/winning/BonusNumberTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import constants.ErrorMessage;
 import java.util.List;
-import lotto.LottoNumber;
+import lotto.core.LottoNumber;
 import org.junit.jupiter.api.Test;
 
 public class BonusNumberTest {
